reorganize some tests.
